everything is working nearly perfecting just want to do sessionStorage now for tools
