+++ frontend/src/contexts/ChatContext.tsx
@@ -56,6 +56,8 @@ export default function ChatProvider({ children }: IContextProvider) {
         setActions,
         logs,
         setLogs,
+        tools,
+        setTools,
         selectedImage,
         setSelectedImage,
         selectedDocument,
@@ -198,6 +200,7 @@ export default function ChatProvider({ children }: IContextProvider) {
                     files,
                     actions,
                     logs,
+                    tools,
                     done,
                     selectedDocument,
                     csvContent,
@@ -210,6 +213,7 @@ export default function ChatProvider({ children }: IContextProvider) {
                     setFiles,
                     setActions,
                     setLogs,
+                    setTools,
                     resetChat,
                     setChats,
                     setMessages,
@@ -238,6 +242,7 @@ export default function ChatProvider({ children }: IContextProvider) {
                 expand,
                 actions,
                 logs,
+                tools,
                 userInput,
                 chatboxRef,
                 chatInputRef,
